requirements.txt
python-dotenv==1.0.0
openai==0.27.8
tiktoken==0.4.0
gradio==3.39.0
gradio_client==0.3.0